global refactoring + k8s
