+++ api/__init__.py
@@ -1,0 +1,1 @@
+"""Инициализация API модулей."""
