+++ k8s/base/kustomization.yaml
@@ -1,0 +1,23 @@
+apiVersion: kustomize.config.k8s.io/v1beta1
+kind: Kustomization
+
+metadata:
+  name: postic-ml-base
+  namespace: postic-ml
+
+resources:
+  - namespace.yaml
+  - persistent-volumes.yaml
+  - redis.yaml
+  - qdrant.yaml
+  - ollama.yaml
+  - app.yaml
+
+labels:
+  - pairs:
+      project: postic-ml
+      version: v1.0.0
+
+images:
+  - name: blackHATred/postic-ml
+    newTag: k8s-latest
