+++ k8s/base/namespace.yaml
@@ -1,0 +1,7 @@
+apiVersion: v1
+kind: Namespace
+metadata:
+  name: postic-ml
+  labels:
+    name: postic-ml
+    purpose: ai-social-media-management
