+++ k8s/overlays/minikube/app-deployment-patch.yaml
@@ -1,0 +1,18 @@
+apiVersion: apps/v1
+kind: Deployment
+metadata:
+  name: postic-ml-app
+  namespace: postic-ml
+spec:
+  replicas: 1  # Уменьшаем количество реплик для Minikube
+  template:
+    spec:
+      containers:
+      - name: postic-ml-app
+        resources:
+          requests:
+            memory: "512Mi"
+            cpu: "250m"
+          limits:
+            memory: "1Gi"
+            cpu: "500m"
