+++ k8s/overlays/minikube/app-service-patch.yaml
@@ -1,0 +1,12 @@
+apiVersion: v1
+kind: Service
+metadata:
+  name: postic-ml-service
+  namespace: postic-ml
+spec:
+  type: NodePort  # Используем NodePort для доступа в Minikube
+  ports:
+  - port: 8000
+    targetPort: 8000
+    nodePort: 30080
+    name: http
