+++ k8s/overlays/minikube/kustomization.yaml
@@ -1,0 +1,32 @@
+apiVersion: kustomize.config.k8s.io/v1beta1
+kind: Kustomization
+
+metadata:
+  name: postic-ml-minikube
+
+resources:
+  - ../../base
+
+patches:
+  - path: ollama-minikube-patch.yaml
+    target:
+      kind: Deployment
+      name: ollama
+  - path: app-deployment-patch.yaml
+    target:
+      kind: Deployment
+      name: postic-ml-app
+  - path: app-service-patch.yaml
+    target:
+      kind: Service
+      name: postic-ml-service
+
+configMapGenerator:
+  - name: postic-ml-config
+    behavior: create
+    literals:
+      - OLLAMA_HOST=http://ollama-service.postic-ml.svc.cluster.local:11434
+      - QDRANT_HOST=qdrant-service.postic-ml.svc.cluster.local
+      - REDIS_HOST=redis-service.postic-ml.svc.cluster.local
+
+namespace: postic-ml
