+++ k8s/overlays/minikube/ollama-minikube-patch.yaml
@@ -1,0 +1,49 @@
+apiVersion: apps/v1
+kind: Deployment
+metadata:
+  name: ollama
+  namespace: postic-ml
+spec:
+  template:
+    spec:
+      # Требуем узел с GPU
+      nodeSelector:
+        accelerator: nvidia-tesla-gpu
+      # Добавляем tolerations для GPU
+      tolerations:
+      - key: nvidia.com/gpu
+        operator: Exists
+        effect: NoSchedule
+      containers:
+      - name: ollama
+        resources:
+          requests:
+            memory: "4Gi"
+            cpu: "1"
+            # Требуем GPU для работы
+            nvidia.com/gpu: 1
+          limits:
+            memory: "8Gi" 
+            cpu: "2"
+            nvidia.com/gpu: 1
+        # Переменные окружения для GPU режима
+        env:
+        - name: OLLAMA_HOST
+          value: "0.0.0.0"
+        - name: OLLAMA_PORT
+          value: "11434"
+        # Включаем GPU
+        - name: NVIDIA_VISIBLE_DEVICES
+          value: "all"
+        - name: NVIDIA_DRIVER_CAPABILITIES
+          value: "compute,utility"
+        # Оптимизация для GPU
+        - name: OLLAMA_NUM_PARALLEL
+          value: "4"
+        - name: OLLAMA_MAX_LOADED_MODELS
+          value: "2"
+        # Принудительно используем GPU
+        - name: CUDA_VISIBLE_DEVICES
+          value: "0"
+        - name: OLLAMA_LLM_LIBRARY
+          value: "cuda"
