+++ k8s/overlays/registry/app-deployment-patch.yaml
@@ -1,0 +1,20 @@
+apiVersion: apps/v1
+kind: Deployment
+metadata:
+  name: postic-ml-app
+  namespace: postic-ml
+spec:
+  replicas: 2  # Больше реплик для продакшена
+  template:
+    spec:
+      containers:
+      - name: postic-ml-app
+        # Политика загрузки образа
+        imagePullPolicy: Always
+        resources:
+          requests:
+            memory: "1Gi"
+            cpu: "500m"
+          limits:
+            memory: "2Gi"
+            cpu: "1"
